+++ src/pdf-backend.c
@@ -57,7 +57,7 @@ Rcairo_backend *Rcairo_new_pdf_backend(Rcairo_backend *be, int conn, const char 
 	be->save_page = pdf_save_page;
 	if (filename){
 		char *fn = NULL;
-		int len = strlen(filename);
+		size_t len = strlen(filename);
 
 		/* Add .pdf extension if necessary */
 		if (len>3 && strcmp(filename+len-4,".pdf")){
